Task completed. Perfomance test added
