+++ TestProject/Models/FilterPropertiesAttribute.cs
@@ -9,25 +9,34 @@ using TestProject.Helpers;
 
 namespace TestProject.Models
 {
-
     [AttributeUsage(AttributeTargets.Method)]
     public sealed class FilterFields : ActionFilterAttribute
     {
         private const string DesiredFieldsParameterFromQueryString = "fields";
 
-        private HashSet<string> _desiredFields;
-        private Type _returnType;
+        private readonly CachableSerializationFilterDecorator _filterDecorator;
 
+        private HashSet<string> _requestedProperties;
+        private Type _returnType;
         private Task _actionExecutionTask;
 
+        public FilterFields()
+        {
+            _filterDecorator = new CachableSerializationFilterDecorator();
+        }
+
+        // if not in cache:
+        // Action executed in 7.0007ms
+        // if cached
+        // Action executed in 6.0006ms
         public override void OnActionExecuted(HttpActionExecutedContext actionExecutedContext)
         {
             var content = actionExecutedContext.Response.Content as ObjectContent;
             if (content != null)
             {
                 WaitForActionExecutingTask();
-                //var contentValue = SerializationDecoratorManager.PrepareDecorator(content.Value, _returnType, _desiredFields);
-                //actionExecutedContext.Response.Content = new ObjectContent(contentValue.GetDecoratorType(), contentValue, content.Formatter);
+                var contentValue = _filterDecorator.GetDecorator(content.Value, _returnType, _requestedProperties);
+                actionExecutedContext.Response.Content = new ObjectContent(contentValue.GetDecoratorType(), contentValue, content.Formatter);
             }
             else
             {
@@ -35,14 +44,18 @@ namespace TestProject.Models
             }
         }
 
+        // if not in cache:
+        // Action executing 64.0064ms
+        // if cached
+        // Action executing 3.0003ms
         public override void OnActionExecuting(HttpActionContext actionContext)
         {
             _actionExecutionTask =
                 Task.Factory.StartNew(() =>
                 {
-                    ParseFieldsNamesFromQueryString(actionContext.Request.RequestUri.Query);
+                    _requestedProperties = ParseFieldsNamesFromQueryString(actionContext.Request.RequestUri.Query);
                     _returnType = actionContext.ActionDescriptor.ReturnType;
-                    //PropertiesFilter.PrepareDelegate(_returnType, _desiredFields);
+                    _filterDecorator.PrepareDecorator(_returnType, _requestedProperties);
                 })
                     .ContinueWith(prevTask =>
                     {
@@ -50,15 +63,16 @@ namespace TestProject.Models
                     });
         }
 
-        private void ParseFieldsNamesFromQueryString(string queryString)
+        private HashSet<string> ParseFieldsNamesFromQueryString(string queryString)
         {
             var param = HttpUtility.ParseQueryString(queryString);
             if (param[DesiredFieldsParameterFromQueryString] != null)
             {
                 var queryParams = param[DesiredFieldsParameterFromQueryString].Split(',');
                 Array.Sort(queryParams);
-                _desiredFields = new HashSet<string>(queryParams);
+                return new HashSet<string>(queryParams);
             }
+            return null;
         }
 
         private void WaitForActionExecutingTask()
